requirements.dev.txt
agate==1.7.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.14.0
blinker==1.7.0
bokeh==2.4.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
dbt-core==1.7.11
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
dbt-snowflake==1.7.3
distro==1.9.0
fastapi==0.110.0
filelock==3.13.3
fonttools==4.49.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
graphdatascience==1.9
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==6.11.0
isodate==0.6.1
jaraco.classes==3.3.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.1
kiwisolver==1.4.5
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.23
leather==0.4.0
Logbook==1.5.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mashumaro==3.12
matplotlib==3.8.3
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
multimethod==1.11.2
mypy-extensions==1.0.0
neo4j==5.18.0
networkx==3.2.1
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pandas==2.2.1
parsedatetime==2.6
pathspec==0.11.2
pillow==10.2.0
platformdirs==3.11.0
protobuf==4.25.3
pyarrow==14.0.2
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.7.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
sqlparse==0.4.4
starlette==0.36.3
streamlit==1.32.0
streamlit-agraph==0.0.45
streamlit-chat==0.1.1
tenacity==8.2.3
text-unidecode==1.3
textdistance==4.6.1
threadpoolctl==3.3.0
tiktoken==0.6.0
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.28.0
wordcloud==1.9.3
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.18.1
